requirements.txt
Flask==1.0.2
tensorflow==1.12.0
Werkzeug==1.0.1
scikit_image==0.13.1
Keras==2.2.4
Pillow==8.2.0
h5py==2.10.0
numpy==1.15.2
